==> Codigo/6. Clustering/2. Cluster_PreparaDatos_DemandaMensual.ipynb <==
## 2. Cluster_PreparaDatos_DemandaMensual

### Objetivo

Proceso general de preparación de datos en donde se identificarán las variables que participarán en los procesos posteriores de clusterización. Se hace énfasis en la variable DEMANDA de viajes consolidado se abrirá para cada uno de los meses de funcionamiento.

### Descripción General de notebook

    1. Carga de datos base:
        - Movimientos históricos
        - Demografía por barrio
        - Geografía por estación Bicimad
        
    2. Concatenación de datasets
    3. Exportar DataFrame a CSV

## 1. Carga de datos base

import time
import pandas as pd

from sklearn.cluster import KMeans
from sklearn.decomposition import PCA
from sklearn.metrics import silhouette_score, adjusted_rand_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import ColumnTransformer

%run "../1. Librerias Mongo/MongoDB_Connections.ipynb"
%run "../1. Librerias Mongo/MongoDB_Funciones_Consultas.ipynb"

### Movimientos históricos

Carga de 2.912.138 registros históricos de viajes Bicimad

t_ini = time.time()
data_Tracks = pd.read_csv('../../Data/DataFrame_Final_Cierre_2017_2019.csv', parse_dates=['FECHA'])
t_end = time.time()
print (t_end - t_ini)

data_Tracks.shape

data_Tracks.info()

data_Tracks.groupby('ESTACION').agg(DEMANDA=('DEMANDA', 'sum'), CUENTA=('DEMANDA', 'count'))

data_Tracks.groupby('ANIO').agg(DEMANDA=('DEMANDA', 'sum'))

### Datos Demográficos por Barrio de Madrid

Los datos demográficos se encuentran cargados dentro de la colección "Demografia" en la base de datos BiciMAD en MongoDB Atlas

# DEMOGRAFIA
db_DemoG = _connect_mongo('cloud', 'cluster0.15npsxw.mongodb.net', None, 'ucmtfm2022', 'UCM_2022', 'BiciMAD', 'Demografia')
data_Demografia = _consulta_Demografia(db_DemoG)
data_Demografia

data_Demografia.info()

### Datos geográficos por Estación BiciMad

Los datos de ubicación geográfica de cada estación de BiciMad se encuentran cargados dentro de la colección "Estaciones" en la base de datos BiciMAD en MongoDB Atlas

db_Stations = _connect_mongo('cloud', 'cluster0.15npsxw.mongodb.net', None, 'ucmtfm2022', 'UCM_2022', 'BiciMAD', 'Estaciones')
data_Stations = _consulta_stations(db_Stations)

data_Stations = data_Stations.drop(columns='Finca')

data_Stations

## 2. Concatenación (Join) de datasets

### Estaciones - Demografia

df_EstacionDemog = pd.merge(data_Stations, data_Demografia, how='left', 
                            left_on=['Distrito', 'Barrio'], 
                            right_on=['Distrito', 'Barrio'])

df_EstacionDemog = df_EstacionDemog.drop(columns=['Distrito_Nombre_y', 'Barrio_Nombre_y'])
df_EstacionDemog.rename(columns = {'Id_Estacion':'ESTACION', 'Distrito_Nombre_x':'Distrito_Nombre', 
                                   'Barrio_Nombre_x':'Barrio_Nombre'}, inplace = True)

df_EstacionDemog

df_EstacionDemog.info()

### Estacion y movimientos consolidados totales por MES

El objetivo de esta concatenación es agregarle al dataframe de Estaciones los siguientes valores representativo de la demanda:

    - DEMANDA_RATIO = ( DEMANDA_TOTAL / MESES ): valor que permita comparar la demanda general entre estaciones, independiente de si tienen 24 meses de funcionamiento o 12 meses de entrada en vigencia.
    
    - DMNDA_yyyymm = Suma de demanda mensual histórica de viajes realizado desde cada estación por cada MES desde Marzo/2017 hasta Diciembre/2020. 




# Creación de DF inicial con la lista de estaciones

estaciones = data_Tracks.ESTACION.unique()
estaciones = pd.DataFrame(estaciones)
estaciones = estaciones.rename(columns={0:'ESTACION'})
estaciones

# Cálculo de meses consolidados totales para cada una de las estaciones

data_Tracks['ANIOMES'] = data_Tracks['ANIO'].astype(str)+("0"+data_Tracks['MES'].astype(str)).str[-2:]

estacion_meses= data_Tracks.groupby('ESTACION').ANIOMES.nunique()
estacion_meses

# Demanda TOTAL por estacion

df_DMNDA_TOTAL = data_Tracks.groupby('ESTACION').agg(DEMANDA=('DEMANDA', 'sum'))
estacion = pd.merge(df_DMNDA_TOTAL, estacion_meses, how='left', left_on='ESTACION', right_on='ESTACION')
estacion['DEMANDA_RATIO'] = estacion['DEMANDA']/estacion['ANIOMES']
estaciones_tracks = pd.merge(estaciones, estacion['DEMANDA_RATIO'], how='left', left_on='ESTACION', right_on='ESTACION')

estaciones_tracks

#### Creación de variables de demanda mensual por estación: df_DMNDA_yyyymm

df_DMNDA_201703 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==3)]
df_DMNDA_201703 = df_DMNDA_201703.groupby('ESTACION').agg(DMNDA_201703=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201703, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201704 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==4)]
df_DMNDA_201704 = df_DMNDA_201704.groupby('ESTACION').agg(DMNDA_201704=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201704, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201705 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==5)]
df_DMNDA_201705 = df_DMNDA_201705.groupby('ESTACION').agg(DMNDA_201705=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201705, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201706 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==6)]
df_DMNDA_201706 = df_DMNDA_201706.groupby('ESTACION').agg(DMNDA_201706=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201706, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201707 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==7)]
df_DMNDA_201707 = df_DMNDA_201707.groupby('ESTACION').agg(DMNDA_201707=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201707, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201708 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==8)]
df_DMNDA_201708 = df_DMNDA_201708.groupby('ESTACION').agg(DMNDA_201708=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201708, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201709 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==9)]
df_DMNDA_201709 = df_DMNDA_201709.groupby('ESTACION').agg(DMNDA_201709=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201709, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201710 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==10)]
df_DMNDA_201710 = df_DMNDA_201710.groupby('ESTACION').agg(DMNDA_201710=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201710, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201711 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==11)]
df_DMNDA_201711 = df_DMNDA_201711.groupby('ESTACION').agg(DMNDA_201711=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201711, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201712 = data_Tracks[(data_Tracks['ANIO']==2017) & (data_Tracks['MES']==12)]
df_DMNDA_201712 = df_DMNDA_201712.groupby('ESTACION').agg(DMNDA_201712=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201712, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201801 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==1)]
df_DMNDA_201801 = df_DMNDA_201801.groupby('ESTACION').agg(DMNDA_201801=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201801, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201802 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==2)]
df_DMNDA_201802 = df_DMNDA_201802.groupby('ESTACION').agg(DMNDA_201802=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201802, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201803 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==3)]
df_DMNDA_201803 = df_DMNDA_201803.groupby('ESTACION').agg(DMNDA_201803=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201803, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201804 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==4)]
df_DMNDA_201804 = df_DMNDA_201804.groupby('ESTACION').agg(DMNDA_201804=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201804, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201805 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==5)]
df_DMNDA_201805 = df_DMNDA_201805.groupby('ESTACION').agg(DMNDA_201805=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201805, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201806 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==6)]
df_DMNDA_201806 = df_DMNDA_201806.groupby('ESTACION').agg(DMNDA_201806=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201806, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201807 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==7)]
df_DMNDA_201807 = df_DMNDA_201807.groupby('ESTACION').agg(DMNDA_201807=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201807, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201808 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==8)]
df_DMNDA_201808 = df_DMNDA_201808.groupby('ESTACION').agg(DMNDA_201808=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201808, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201809 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==9)]
df_DMNDA_201809 = df_DMNDA_201809.groupby('ESTACION').agg(DMNDA_201809=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201809, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201810 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==10)]
df_DMNDA_201810 = df_DMNDA_201810.groupby('ESTACION').agg(DMNDA_201810=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201810, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201811 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==11)]
df_DMNDA_201811 = df_DMNDA_201811.groupby('ESTACION').agg(DMNDA_201811=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201811, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201812 = data_Tracks[(data_Tracks['ANIO']==2018) & (data_Tracks['MES']==12)]
df_DMNDA_201812 = df_DMNDA_201812.groupby('ESTACION').agg(DMNDA_201812=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201812, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201901 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==1)]
df_DMNDA_201901 = df_DMNDA_201901.groupby('ESTACION').agg(DMNDA_201901=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201901, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201902 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==2)]
df_DMNDA_201902 = df_DMNDA_201902.groupby('ESTACION').agg(DMNDA_201902=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201902, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201903 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==3)]
df_DMNDA_201903 = df_DMNDA_201903.groupby('ESTACION').agg(DMNDA_201903=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201903, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201904 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==4)]
df_DMNDA_201904 = df_DMNDA_201904.groupby('ESTACION').agg(DMNDA_201904=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201904, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201905 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==5)]
df_DMNDA_201905 = df_DMNDA_201905.groupby('ESTACION').agg(DMNDA_201905=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201905, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201906 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==6)]
df_DMNDA_201906 = df_DMNDA_201906.groupby('ESTACION').agg(DMNDA_201906=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201906, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201907 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==7)]
df_DMNDA_201907 = df_DMNDA_201907.groupby('ESTACION').agg(DMNDA_201907=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201907, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201908 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==8)]
df_DMNDA_201908 = df_DMNDA_201908.groupby('ESTACION').agg(DMNDA_201908=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201908, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201909 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==9)]
df_DMNDA_201909 = df_DMNDA_201909.groupby('ESTACION').agg(DMNDA_201909=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201909, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201910 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==10)]
df_DMNDA_201910 = df_DMNDA_201910.groupby('ESTACION').agg(DMNDA_201910=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201910, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201911 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==11)]
df_DMNDA_201911 = df_DMNDA_201911.groupby('ESTACION').agg(DMNDA_201911=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201911, how='left', left_on='ESTACION', right_on='ESTACION')

df_DMNDA_201912 = data_Tracks[(data_Tracks['ANIO']==2019) & (data_Tracks['MES']==12)]
df_DMNDA_201912 = df_DMNDA_201912.groupby('ESTACION').agg(DMNDA_201912=('DEMANDA', 'sum'))
estaciones_tracks = pd.merge(estaciones_tracks, df_DMNDA_201912, how='left', left_on='ESTACION', right_on='ESTACION')

#### El dataframe final contiene las variables DEMANDA_RATIO y DMNDA_yyyymm para cada una de las estaciones

estaciones_tracks.fillna(0, inplace=True)
estaciones_tracks

data_Tracks['ANIOMES'] = data_Tracks['ANIO'].astype(str)+("0"+data_Tracks['MES'].astype(str)).str[-2:]
data_Tracks

estaciones_tracks.info()

### Concatenación de Estaciones con Datos Demográficos por Barrio

df_FinalCluster = pd.merge(estaciones_tracks, df_EstacionDemog, how='left', left_on='ESTACION', right_on='ESTACION')
df_FinalCluster = df_FinalCluster.drop(columns=['Nro_Estacion', 'Fec_Alta', 'Distrito_Nombre', 'Barrio_Nombre', 
                                                'Calle', 'Direccion', 'Tipo_Reserva'])
df_FinalCluster


df_FinalCluster.info()

# Cambios de tipos de datos para variables numéricas

df_FinalCluster['Gis_X'] = df_FinalCluster.Gis_X.astype(str)
df_FinalCluster['Gis_X'] = [x.replace(',', '.') for x in df_FinalCluster['Gis_X']]
df_FinalCluster['Gis_X'] = df_FinalCluster.Gis_X.astype(float)

df_FinalCluster['Gis_Y'] = df_FinalCluster.Gis_Y.astype(str)
df_FinalCluster['Gis_Y'] = [x.replace(',', '.') for x in df_FinalCluster['Gis_Y']]
df_FinalCluster['Gis_Y'] = df_FinalCluster.Gis_Y.astype(float)

#df_FinalCluster['Finca'] = df_FinalCluster.Finca.astype('string')
df_FinalCluster['Barrio'] = df_FinalCluster.Barrio.astype('string')

df_FinalCluster['Tasa_Paro'] = df_FinalCluster.Tasa_Paro.astype(str)
df_FinalCluster['Tasa_Paro'] = [x.replace(',', '.') for x in df_FinalCluster['Tasa_Paro']]
df_FinalCluster['Tasa_Paro'] = df_FinalCluster.Tasa_Paro.astype(float)

df_FinalCluster['Renta_Media_Persona'] = df_FinalCluster.Renta_Media_Persona.astype(str)
df_FinalCluster['Renta_Media_Persona'] = [x.replace(',', '.') for x in df_FinalCluster['Renta_Media_Persona']]
df_FinalCluster['Renta_Media_Persona'] = df_FinalCluster.Renta_Media_Persona.astype(float)

df_FinalCluster['Renta_Media_Hogar'] = df_FinalCluster.Renta_Media_Hogar.astype(str)
df_FinalCluster['Renta_Media_Hogar'] = [x.replace(',', '.') for x in df_FinalCluster['Renta_Media_Hogar']]
df_FinalCluster['Renta_Media_Hogar'] = df_FinalCluster.Renta_Media_Hogar.astype(float)

df_FinalCluster['Gis_X'] = df_FinalCluster.Gis_X.astype(str)
df_FinalCluster['Gis_X'] = [x.replace(',', '.') for x in df_FinalCluster['Gis_X']]
df_FinalCluster['Gis_X'] = df_FinalCluster.Gis_X.astype(float)

df_FinalCluster['Gis_Y'] = df_FinalCluster.Gis_Y.astype(str)
df_FinalCluster['Gis_Y'] = [x.replace(',', '.') for x in df_FinalCluster['Gis_Y']]
df_FinalCluster['Gis_Y'] = df_FinalCluster.Gis_Y.astype(float)


df_FinalCluster.info()
df_FinalCluster


## 3. Exportar DataFrame a CSV

df_FinalCluster.to_csv('../../data/df_FinalCluster_2017_2019.csv', index=False)
