==> Codigo/6. Clustering/6. Cluster_SoloDemandaPorMes.ipynb <==
## 6. Cluster_DatosFull

### Objetivo

Generar cluster's de estaciones en base sólo a la variable de demanda histórica

### Descripción General de notebook

    1. Carga de datos base proveniente de etapa de preparación
    2. Clusterización final en base a k=5

## 1. Carga de datos de entrada

import pandas as pd
from sklearn.cluster import KMeans
import sklearn.cluster as cluster
from sklearn import preprocessing
from sklearn.pipeline import make_pipeline
import sklearn.cluster as cluster
import sklearn.metrics as metrics
from sklearn.preprocessing import OneHotEncoder
import pandas as pd
from sklearn.preprocessing import MinMaxScaler
import seaborn as sns
from matplotlib import pyplot as plt
from sklearn.decomposition import PCA
import numpy as np

%matplotlib inline

pd.set_option('display.max_rows', 5)

df_FinalCluster_Original = pd.read_csv('../../data/df_FinalCluster_2017_2019.csv',
                              dtype={'Distrito':'category','Plazas':'category','Barrio':'category'})

df_FinalCluster_Original.info()
#df_FinalCluster

### Selección de Variables DEMANDA Mensual

df_FinalCluster = df_FinalCluster_Original.copy()

df_FinalCluster = df_FinalCluster.drop(columns='ESTACION')

df_FinalCluster = df_FinalCluster[[
    'DMNDA_201703', 'DMNDA_201704', 'DMNDA_201705', 'DMNDA_201706', 
    'DMNDA_201707', 'DMNDA_201708', 'DMNDA_201709', 'DMNDA_201710', 'DMNDA_201711', 'DMNDA_201712',
    'DMNDA_201801', 'DMNDA_201802', 'DMNDA_201803', 'DMNDA_201804', 'DMNDA_201805', 'DMNDA_201806', 
    'DMNDA_201807', 'DMNDA_201808', 'DMNDA_201809', 'DMNDA_201810', 'DMNDA_201811', 'DMNDA_201812',
    'DMNDA_201901', 'DMNDA_201902', 'DMNDA_201903', 'DMNDA_201904', 'DMNDA_201905', 'DMNDA_201906', 
    'DMNDA_201907', 'DMNDA_201908', 'DMNDA_201909', 'DMNDA_201910', 'DMNDA_201911', 'DMNDA_201912']]

df_FinalCluster.info()

num_cols = list(df_FinalCluster._get_numeric_data().columns)
cat_cols = list(set(df_FinalCluster.columns) - set(num_cols))

df_FinalCluster

print(num_cols)
print(cat_cols)

### Concatenación de DF: Numéricos y Categóricas

df_knn_final = df_FinalCluster[num_cols].copy()
df_knn_final.info()

## 2. Clusterización final en base a k=5

La clusterización se ejecuta para cada DataFrame según PCA siempre considerando k=5. 
La ejecución del algoritmo final considera:

    - Estandarización: MinMaxScaler()
    - Clusterización: KMeans según parámetros observados en siguiente celda

n_cluster = 5

scaler = MinMaxScaler()
kmeans = KMeans(
               n_clusters=n_cluster,
               init="k-means++",
               n_init=50,
               max_iter=500,
               random_state=200,)

pipeline = make_pipeline(scaler, kmeans)
pipeline

df_pred_knn_final_t = pipeline.fit_transform(df_knn_final)
df_pred_knn_final = pipeline.fit_predict(df_knn_final)


df_pred_knn_final_t

df_pred_knn_final

df_final_Clusterizado = pd.concat([df_FinalCluster_Original, pd.DataFrame(df_pred_knn_final)], axis=1)

df_final_Clusterizado.rename(columns={0: 'CLUSTER'}, inplace=True)
df_final_Clusterizado.info()

df_final_Clusterizado[['ESTACION', 'CLUSTER']]

df_final_Clusterizado[['ESTACION', 'CLUSTER']].to_csv('../../data/Clusterizada_soloDemandaMensual.csv', index=False)
