==> Codigo/6. Clustering/8. ConcatenacionDF_Final.ipynb <==
## 8. ConcatenacionDF_Final

### Objetivo

Generar dataframe final de movimientos a utilizar en etapas de machine learning posteriores. Este DF contendrá la identificación del cluster al cual pertenece cada estación según tipo de clusterización (demanda, geografica o full)

### Descripción General de notebook

    1. Carga de datos de movimientos históricos y de clusterización
    2. Concatenación de dataframes
    3. Export de datos finales

## 1. Carga de datos de entrada

import pandas as pd
import numpy as np
import time
import datetime

%run "../1. Librerias Mongo/MongoDB_Connections.ipynb"
%run "../1. Librerias Mongo/MongoDB_Funciones_Consultas.ipynb"

# MOVIMIENTOS

df_tracks = pd.read_csv('../../data/DataFrame_Final_Cierre_2017_2019.csv', parse_dates=['FECHA'])

# RESULTADOS DE CLUSTERIZACION
df_clusterFull = pd.read_csv('../../data/Clusterizada_DataFull.csv')
df_clusterGeog5 = pd.read_csv('../../data/Clusterizada_soloGeografico_5Cluster.csv')
df_clusterGeog7 = pd.read_csv('../../data/Clusterizada_soloGeografico_7Cluster.csv')
df_clusterDmnd = pd.read_csv('../../data/Clusterizada_soloDemandaMensual.csv')

df_clusterDmnd

## 2. Concatenación de dataframes

df_final = pd.merge(df_tracks, df_clusterFull[['ESTACION','CLUSTER']], how='left', left_on=['ESTACION'], right_on=['ESTACION'])
df_final = df_final.rename(columns={'CLUSTER': 'CLUSTER_FullCols'})
df_final

df_final = pd.merge(df_final, df_clusterGeog5[['ESTACION','CLUSTER']], how='left', left_on=['ESTACION'], right_on=['ESTACION'])
df_final = df_final.rename(columns={'CLUSTER': 'CLUSTER_soloGeo5'})
df_final

df_final = pd.merge(df_final, df_clusterGeog7[['ESTACION','CLUSTER']], how='left', left_on=['ESTACION'], right_on=['ESTACION'])
df_final = df_final.rename(columns={'CLUSTER': 'CLUSTER_soloGeo7'})
df_final

df_final = pd.merge(df_final, df_clusterDmnd[['ESTACION','CLUSTER']], how='left', left_on=['ESTACION'], right_on=['ESTACION'])
df_final = df_final.rename(columns={'CLUSTER': 'CLUSTER_soloDemanda'})
df_final

## 3. Export de datos finales

df_final.info()

# REORDENAMIENTO DE COLUMNAS PREVIO A EXPORT

df_final_cierre = df_final[['ESTACION', 'ANIO', 'MES', 'DIA', 'HORA', 'FECHA', 'DIA_SEMANA', 'AM_PM', 'TEMPORADA', 
                            'TEMPORADA_NUM', 'Es_Festivo', 'Es_FinSemana', 'TEMPERATURA', 'VIENTO', 'PRESION', 'HUMEDAD', 
                            'PRECIPITACION_1h', 'PRECIPITACION_3h', 'DESC_TIEMPO', 'DESC_TIEMPO_detalle', 'CLUSTER_FullCols', 
                            'CLUSTER_soloGeo5', 'CLUSTER_soloGeo7', 'CLUSTER_soloDemanda', 'DEMANDA']]
df_final_cierre

df_final_cierre.info()

df_final_cierre.to_csv('../../data/DataFrame_Final_Cierre_Cluster_2017_2019.csv', index=False)

df_final_cierre.groupby(['ANIO','MES']).agg(DEMANDA=('DEMANDA', 'sum'))
